requirements.txt
matplotlib>=3.0
pillow>=9.0

